analyzer: Allow for data to be provided
If the user has the file content in a buffer it is more efficient to be
able to hand it to the analyzer directly instead of writing it out into
a temporary file just to be read back in again by the requests module.
Add an observable key 'dataProvided' which can be set to True to change
behaviour of run_by_id() and in turn run_by_name() so it uses the 'data'
key as observable data directly instead of interpreting it as a file
name and opening that file.

Signed-off-by: Michael Weiser <michael.weiser@gmx.de>